+++ CHANGES.rst
@@ -1,9 +1,15 @@
 .. :changelog:
 
 History
--------
+=======
+
+0.2.0
+-----
+* add RabbitMQ authentication
+* add Django 1.11 and Python 3.6 support
+
 
 0.1.0 (2017-03-10)
-++++++++++++++++++
+------------------
 
 * First release on PyPI.